+++ src/main/java/zaftnotameni/creatania/CreataniaMain.java
@@ -1,16 +1,13 @@
 package zaftnotameni.creatania;
 
-import com.simibubi.create.foundation.data.LangMerger;
-import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
 import zaftnotameni.creatania.event.ForgeEventBus;
 import zaftnotameni.creatania.registry.*;
-import zaftnotameni.creatania.registry.datagen.LangPartials;
+import zaftnotameni.creatania.registry.datagen.ForgeDatagenInitializer;
 import zaftnotameni.creatania.setup.*;
 
 import static zaftnotameni.creatania.Constants.MODID;
@@ -45,23 +42,7 @@ public class CreataniaMain {
     Blocks.register(bus);
     Fluids.register(bus);
     BlockEntities.register(bus);
-    Index.CREATE_REGISTRATE.registerEventListeners(bus);
     Recipes.register(bus);
-    bus.addListener(EventPriority.LOWEST, CreataniaMain::gatherData);
-  }
-
-  public static void gatherData(GatherDataEvent event) {
-    // TagGen.datagen();
-    DataGenerator gen = event.getGenerator();
-    if (event.includeClient()) {
-      gen.addProvider(new LangMerger(gen, MODID, "Creatania", LangPartials.values()));
-    }
-    if (event.includeServer()) {
-//      gen.addProvider(new AllAdvancements(gen));
-//      gen.addProvider(new StandardRecipeGen(gen));
-//      gen.addProvider(new MechanicalCraftingRecipeGen(gen));
-//      gen.addProvider(new SequencedAssemblyRecipeGen(gen));
-//      ProcessingRecipeGen.registerAll(gen);
-    }
+    bus.addListener(EventPriority.LOWEST, ForgeDatagenInitializer::configureForgeDatagen);
   }
 }