+++ src/main/java/zaftnotameni/creatania/registry/BlockEntities.java
@@ -1,7 +1,6 @@
 package zaftnotameni.creatania.registry;
 
 import com.simibubi.create.content.contraptions.base.HalfShaftInstance;
-import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.tterrag.registrate.util.entry.BlockEntityEntry;
 import net.minecraftforge.eventbus.api.IEventBus;
 import zaftnotameni.creatania.machines.manacondenser.ManaCondenserBlockEntity;
@@ -19,45 +18,43 @@ import zaftnotameni.creatania.stress.xorlever.XorLeverRenderer;
 import zaftnotameni.creatania.util.Log;
 
 public class BlockEntities {
-  public static final CreateRegistrate CREATE_REGISTRATE = Index.getCreateRegistrate()
-    .creativeModeTab(() -> CreativeModeTabs.CREATANIA_ITEMS);
-
-  public static final BlockEntityEntry<ManaMotorBlockEntity> MANA_MOTOR_BLOCK_ENTITY = CREATE_REGISTRATE
+  
+  public static final BlockEntityEntry<ManaMotorBlockEntity> MANA_MOTOR_BLOCK_ENTITY = Index.all()
     .tileEntity("mana_motor_block_entity", ManaMotorBlockEntity::new)
     .instance(() -> HalfShaftInstance::new)
     .validBlocks(Blocks.MANA_MOTOR)
     .renderer(() -> ManaMotorRenderer::new)
     .register();
 
-  public static final BlockEntityEntry<ManaGeneratorBlockEntity> MANA_GENERATOR_BLOCK_ENTITY = CREATE_REGISTRATE
+  public static final BlockEntityEntry<ManaGeneratorBlockEntity> MANA_GENERATOR_BLOCK_ENTITY = Index.all()
     .tileEntity("mana_generator_block_entity", ManaGeneratorBlockEntity::new)
     .instance(() -> HalfShaftInstance::new)
     .validBlocks(Blocks.MANA_GENERATOR)
     .renderer(() -> ManaGeneratorRenderer::new)
     .register();
 
-  public static final BlockEntityEntry<ManaCondenserBlockEntity> MANA_CONDENSER_BLOCK_ENTITY = CREATE_REGISTRATE
+  public static final BlockEntityEntry<ManaCondenserBlockEntity> MANA_CONDENSER_BLOCK_ENTITY = Index.all()
     .tileEntity("mana_condenser_block_entity", ManaCondenserBlockEntity::new)
     .instance(() -> HalfShaftInstance::new)
     .validBlocks(Blocks.MANA_CONDENSER)
     .renderer(() -> ManaCondenserRenderer::new)
     .register();
 
-  public static final BlockEntityEntry<OmniboxBlockEntity> OMNIBOX_BLOCK_ENTITY = CREATE_REGISTRATE
+  public static final BlockEntityEntry<OmniboxBlockEntity> OMNIBOX_BLOCK_ENTITY = Index.all()
     .tileEntity("omnibox_block_entity", OmniboxBlockEntity::new)
     .instance(() -> HalfShaftInstance::new)
     .validBlocks(Blocks.OMNIBOX)
     .renderer(() -> OmniboxRenderer::new)
     .register();
 
-  public static final BlockEntityEntry<XorLeverBlockEntity> XOR_LEVER_BLOCK_ENTITY = CREATE_REGISTRATE
+  public static final BlockEntityEntry<XorLeverBlockEntity> XOR_LEVER_BLOCK_ENTITY = Index.all()
     .tileEntity("xor_lever_block_entity", XorLeverBlockEntity::new)
     .instance(() -> XorLeverInstance::new, false)
     .validBlocks(Blocks.XOR_LEVER)
     .renderer(() -> XorLeverRenderer::new)
     .register();
 
-  public static final BlockEntityEntry<BlazuniaFunctionalFlowerBlockEntity> BLAZUNIA_BLOCK_ENTITY = CREATE_REGISTRATE
+  public static final BlockEntityEntry<BlazuniaFunctionalFlowerBlockEntity> BLAZUNIA_BLOCK_ENTITY = Index.all()
     .tileEntity("blazunia_block_entity", BlazuniaFunctionalFlowerBlockEntity::new)
     .validBlocks(Blocks.BLAZUNIA_BLOCK)
     .register();