+++ src/main/java/zaftnotameni/creatania/registry/datagen/ForgeDatagenInitializer.java
@@ -1,5 +1,7 @@
 package zaftnotameni.creatania.registry.datagen;
 import com.simibubi.create.foundation.data.LangMerger;
+import net.minecraft.data.DataGenerator;
+import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
@@ -14,10 +16,14 @@ public class ForgeDatagenInitializer {
     var generator = evt.getGenerator();
     var helper = evt.getExistingFileHelper();
     // var helper = new ExistingFileHelper(Collections.emptyList(), Collections.emptySet(), false, null, null);
-
-
     generator.addProvider(new LangMerger(generator, Constants.MODID, "Creatania", LangPartials.values()));
 
+    if (evt.includeServer()) {
+      serverDataGen(generator, helper);
+    }
+
+  }
+  public static void serverDataGen(DataGenerator generator, ExistingFileHelper helper) {
     var blockTagProvider = new ForgeBlockTagProvider(generator, helper);
 
     generator.addProvider(blockTagProvider);
@@ -29,16 +35,11 @@ public class ForgeDatagenInitializer {
     generator.addProvider(new ForgeSequencedAssemblyRecipeProvider(generator));
     generator.addProvider(new ForgeAdvancementsProvider(generator));
 
-
-    if (evt.includeServer()) {
-      ForgeCreateProcessingRecipeProvider.registerAll(generator);
-      generator.addProvider(new ElvenTradeRecipeGen(generator));
-      generator.addProvider(new ManaInfusionRecipeGen(generator));
-      generator.addProvider(new RuneAltarRecipeGen(generator));
-      generator.addProvider(new TerraPlateRecipeGen(generator));
-      generator.addProvider(new PureDaisyRecipeGen(generator));
-      generator.addProvider(new ForgeLangProvider(generator));
-    }
-
+    ForgeCreateProcessingRecipeProvider.registerAll(generator);
+    generator.addProvider(new ElvenTradeRecipeGen(generator));
+    generator.addProvider(new ManaInfusionRecipeGen(generator));
+    generator.addProvider(new RuneAltarRecipeGen(generator));
+    generator.addProvider(new TerraPlateRecipeGen(generator));
+    generator.addProvider(new PureDaisyRecipeGen(generator));
   }
 }