+++ src/main/java/zaftnotameni/creatania/setup/FMLCommonSetup.java
@@ -1,8 +1,8 @@
 package zaftnotameni.creatania.setup;
 import com.simibubi.create.content.contraptions.fluids.tank.BoilerHeaters;
 import com.simibubi.create.foundation.block.BlockStressValues;
-import com.simibubi.create.foundation.config.AllConfigs;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import zaftnotameni.creatania.config.CommonConfig;
 import zaftnotameni.creatania.network.Networking;
 import zaftnotameni.creatania.registry.Advancements;
 import zaftnotameni.creatania.registry.Blocks;
@@ -13,13 +13,15 @@ import static zaftnotameni.creatania.Constants.MODID;
 public class FMLCommonSetup {
   public static void run(final FMLCommonSetupEvent event) {
     Log.LOGGER.info("creatania setup started");
-    BlockStressValues.registerProvider(MODID, AllConfigs.SERVER.kinetics.stressValues);
+    // BlockStressValues.registerProvider(MODID, AllConfigs.SERVER.kinetics.stressValues);
+    BlockStressValues.registerProvider(MODID, CommonConfig.stressProvider);
     BoilerHeaters.registerHeater(Blocks.REAL_MANA_BLOCK.get(), (level, pos, state) -> 4);
     BoilerHeaters.registerHeater(Blocks.PURE_MANA_BLOCK.get(), (level, pos, state) -> 2);
     Networking.registerMessages();
     event.enqueueWork(() -> {
       Advancements.register();
       Triggers.register();
+      CommonConfig.stressProvider.registerAll();
     });
     Log.LOGGER.info("creatania setup finished");
   }